one_shot_personalization/requirements.txt
torch>=2.2.0
transformers>=4.42.0
accelerate
safetensors
Pillow

diffusers[torch]>=0.29.0

# face & utils
insightface==0.7.3
onnxruntime-gpu
opencv-python

# UI
gradio